+++ src/CleanRegex/Internal/Delimiter/UndelimiterablePatternException.php
@@ -1,0 +1,6 @@
+<?php
+namespace TRegx\CleanRegex\Internal\Delimiter;
+
+class UndelimiterablePatternException extends \Exception
+{
+}